Creating the DB with events
Creating a DB with 3 table: events, users, booking.

I also added 24 events distributed between the 18-23 of December 2023, with each event happening at a specific hour and day. There is no repetition (each event only occurs once).

Also added images for each event just in case we can display them. I added the images because if we went with URL the best practice would be to create a new table that would store the values. If you think it is too confusing this way, I can add that table to the DB.

We need to update the Lat and Long with the coordinates we want.